+++ yash-cli/tests/scripted_test/intr-y.sh
@@ -36,3 +36,10 @@ __IN__
 INT
 The file was not created.
 __OUT__
+
+test_o -e 0 'interrupting a command substitution in interactive mode' -im --norcfile
+echo $(sh -c 'kill -INT $$'); echo 'This should not be printed.'
+kill -l $?
+__IN__
+INT
+__OUT__